requirements.txt
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
build==0.10.0
CacheControl==0.12.14
certifi==2023.7.22
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.6
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.7
dnspython==2.4.1
dulwich==0.21.5
email-validator==2.0.0.post2
filelock==3.12.2
Flask==2.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-RESTful==0.3.10
Flask-WTF==1.1.1
greenlet==2.0.2
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
installer==0.7.0
itsdangerous==2.0.1
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==23.13.1
lockfile==0.12.2
MarkupSafe==2.1.3
mongoengine==0.27.0
more-itertools==10.1.0
msgpack==1.0.5
packaging==23.1
pexpect==4.8.0
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.9.1
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
ptyprocess==0.7.0
pymongo==4.4.1
pyproject_hooks==1.0.0
python-dotenv==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.2
rapidfuzz==2.15.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.9.2
shellingham==1.5.0.post1
six==1.16.0
SQLAlchemy==2.0.19
svgwrite==1.4.3
tomlkit==0.12.1
Tree==0.2.4
trove-classifiers==2023.8.7
typing_extensions==4.7.1
urllib3==1.26.16
virtualenv==20.24.2
webencodings==0.5.1
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.16.2
